+++ frontend/src/App.css
@@ -1,35 +1,7 @@
-.container {
-  max-width: 600px;
-  margin: auto;
-  text-align: center;
-  font-family: Arial, sans-serif;
-}
-
-textarea {
-  width: 100%;
-  height: 100px;
-  margin: 10px 0;
-  padding: 10px;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-}
-
-button {
-  background-color: #007bff;
-  color: white;
-  padding: 10px;
-  border: none;
-  cursor: pointer;
-  border-radius: 5px;
-}
-
-button:disabled {
-  background-color: gray;
-}
-
-.response {
-  margin-top: 20px;
-  padding: 10px;
-  background: #f3f3f3;
-  border-radius: 5px;
+.app-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+  background: #121212;
 }
