+++ frontend/src/TutorSession/TutorSession.css
@@ -1,0 +1,88 @@
+.tutor-session {
+    max-width: 600px;
+    margin: 40px auto;
+    padding: 20px;
+    background: linear-gradient(135deg, #1e3c72, #2a5298);
+    border-radius: 12px;
+    box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.2);
+    text-align: center;
+    color: white;
+  }
+  
+  .title {
+    font-size: 1.8rem;
+    font-weight: bold;
+    margin-bottom: 15px;
+  }
+  
+  .label {
+    font-size: 1rem;
+    display: block;
+    margin: 10px 0;
+  }
+  
+  .dropdown {
+    width: 100%;
+    padding: 10px;
+    border-radius: 6px;
+    border: none;
+    font-size: 1rem;
+    cursor: pointer;
+    background: #f8f8f8;
+    margin-bottom: 15px;
+  }
+  
+  .question-input {
+    width: 100%;
+    height: 80px;
+    padding: 10px;
+    border-radius: 6px;
+    border: none;
+    font-size: 1rem;
+    resize: none;
+    background: white;
+    color: black;
+    margin-bottom: 15px;
+  }
+  
+  .file-upload {
+    width: 100%;
+    padding: 8px;
+    margin: 10px 0;
+    background: white;
+    border-radius: 6px;
+  }
+  
+  .ask-button {
+    background: #ff9800;
+    color: white;
+    padding: 12px 20px;
+    border: none;
+    border-radius: 6px;
+    font-size: 1rem;
+    cursor: pointer;
+    transition: 0.3s;
+  }
+  
+  .ask-button:hover {
+    background: #e68900;
+  }
+  
+  .response-container {
+    margin-top: 20px;
+    padding: 15px;
+    background: rgba(255, 255, 255, 0.1);
+    border-radius: 6px;
+  }
+  
+  .response-title {
+    font-size: 1.4rem;
+    font-weight: bold;
+    margin-bottom: 10px;
+  }
+  
+  .response-text {
+    font-size: 1rem;
+    word-wrap: break-word;
+  }
+
